+++ cr_electronic_invoice/models/__init__.py
@@ -5,5 +5,7 @@ from . import res_partner
 from . import res_company
 from . import electronic_invoice
 from . import account_invoice
+from . import account_invoice_line
+from . import account_invoice_refund
 from . import account_journal
 from . import facturacion_electronica
